Improved GPM::Uninstall command to take into account multiple similar dependencies and output format similar to GPM::Install

+++ system/src/Grav/Console/Gpm/UninstallCommand.php
@@ -39,6 +39,8 @@ class UninstallCommand extends ConsoleCommand
      */
     protected $tmp;
 
+    protected $dependencies= [];
+
     /**
      *
      */
@@ -108,15 +110,12 @@ class UninstallCommand extends ConsoleCommand
                 $this->output->writeln("  '- <red>Installation failed or aborted.</red>");
                 $this->output->writeln('');
             } else {
-                $this->output->write("  |- Uninstalling package...  ");
                 $uninstall = $this->uninstallPackage($slug, $package);
 
                 if (!$uninstall) {
                     $this->output->writeln("  '- <red>Uninstallation failed or aborted.</red>");
-                    $this->output->writeln('');
                 } else {
                     $this->output->writeln("  '- <green>Success!</green>  ");
-                    $this->output->writeln('');
                 }
             }
 
@@ -156,36 +155,29 @@ class UninstallCommand extends ConsoleCommand
             return false;
         }
 
-        $locator = Grav::instance()['locator'];
-        $path = $locator->findResource($package->package_type . '://' . $slug);
-        Installer::uninstall($path);
-        $errorCode = Installer::lastErrorCode();
-
-        if ($errorCode && $errorCode !== Installer::IS_LINK && $errorCode !== Installer::EXISTS) {
-            $this->output->write("\x0D");
-            // extra white spaces to clear out the buffer properly
-            $this->output->writeln("  |- Uninstalling package...  <red>error</red>                             ");
-            $this->output->writeln("  |  '- " . Installer::lastErrorMsg());
-
-            return false;
-        }
-
-        $message = Installer::getMessage();
-        if ($message) {
-            $this->output->write("\x0D");
-            // extra white spaces to clear out the buffer properly
-            $this->output->writeln("  |- " . $message);
-        }
-
-        $this->output->write("\x0D");
-        // extra white spaces to clear out the buffer properly
-        $this->output->writeln("  |- Uninstalling package...  <green>ok</green>                             ");
-
-
         if (isset($package->dependencies)) {
+
+            $dependencies = $package->dependencies;
+
+            if ($is_dependency) {
+                foreach ($dependencies as $key => $dependency) {
+                    if (in_array($dependency['name'], $this->dependencies)) {
+                        unset($dependencies[$key]);
+                    }
+                }
+            } else {
+                if (count($dependencies) > 0) {
+                    $this->output->writeln('  `- Dependencies found...');
+                    $this->output->writeln('');
+                }
+            }
+
             $questionHelper = $this->getHelper('question');
 
-            foreach($package->dependencies as $dependency) {
+            foreach ($dependencies as $dependency) {
+
+                $this->dependencies[] = $dependency['name'];
+
                 if (is_array($dependency)) {
                     $dependency = $dependency['name'];
                 }
@@ -194,24 +186,58 @@ class UninstallCommand extends ConsoleCommand
                 }
 
                 $dependencyPackage = $this->gpm->findPackage($dependency);
-                $question = new ConfirmationQuestion("  |  '- Delete dependency <cyan>" . $dependency . "</cyan> too? [y|N] ", false);
-                $answer = $questionHelper->ask($this->input, $this->output, $question);
 
-                if ($answer) {
-                    $this->output->writeln("  |     '- <red>You decided to delete " . $dependency . ".</red>");
+                $dependency_exists = $this->packageExists($dependency, $dependencyPackage);
 
-                    $uninstall = $this->uninstallPackage($dependency, $dependencyPackage, true);
+                if ($dependency_exists == Installer::EXISTS) {
+                    $this->output->writeln("A dependency on <cyan>" . $dependencyPackage->name . "</cyan> [v" . $dependencyPackage->version . "] was found");
 
-                    if (!$uninstall) {
-                        $this->output->writeln("  '- <red>Uninstallation failed or aborted.</red>");
+                    $question = new ConfirmationQuestion("  |- Uninstall <cyan>" . $dependencyPackage->name . "</cyan>? [y|N] ", false);
+                    $answer = $questionHelper->ask($this->input, $this->output, $question);
+
+                    if ($answer) {
+                        $uninstall = $this->uninstallPackage($dependency, $dependencyPackage, true);
+
+                        if (!$uninstall) {
+                            $this->output->writeln("  '- <red>Uninstallation failed or aborted.</red>");
+                        } else {
+                            $this->output->writeln("  '- <green>Success!</green>  ");
+
+                        }
                         $this->output->writeln('');
                     } else {
-                        $this->output->writeln("  '- <green>Success!</green>  ");
+                        $this->output->writeln("  '- <yellow>You decided not to uninstall " . $dependencyPackage->name . ".</yellow>");
                         $this->output->writeln('');
                     }
                 }
+
             }
         }
+
+
+        $locator = Grav::instance()['locator'];
+        $path = $locator->findResource($package->package_type . '://' . $slug);
+        Installer::uninstall($path);
+        $errorCode = Installer::lastErrorCode();
+
+        if ($errorCode && $errorCode !== Installer::IS_LINK && $errorCode !== Installer::EXISTS) {
+            $this->output->writeln("  |- Uninstalling " . $package->name . " package...  <red>error</red>                             ");
+            $this->output->writeln("  |  '- <yellow>" . Installer::lastErrorMsg()."</yellow>");
+
+            return false;
+        }
+
+        $message = Installer::getMessage();
+        if ($message) {
+            $this->output->writeln("  |- " . $message);
+        }
+
+        if (!$is_dependency && $this->dependencies) {
+            $this->output->writeln("Finishing up uninstalling <cyan>" . $package->name . "</cyan>");
+        }
+        $this->output->writeln("  |- Uninstalling " . $package->name . " package...  <green>ok</green>                             ");
+
+
 
         return true;
     }
@@ -225,13 +251,12 @@ class UninstallCommand extends ConsoleCommand
 
     private function checkDestination($slug, $package)
     {
-        $path = Grav::instance()['locator']->findResource($package->package_type . '://' . $slug);
         $questionHelper = $this->getHelper('question');
         $skipPrompt = $this->input->getOption('all-yes');
 
-        Installer::isValidDestination($path);
+        $exists = $this->packageExists($slug, $package);
 
-        if (Installer::lastErrorCode() == Installer::IS_LINK) {
+        if ($exists == Installer::IS_LINK) {
             $this->output->write("\x0D");
             $this->output->writeln("  |- Checking destination...  <yellow>symbolic link</yellow>");
 
@@ -246,7 +271,7 @@ class UninstallCommand extends ConsoleCommand
             $answer = $questionHelper->ask($this->input, $this->output, $question);
 
             if (!$answer) {
-                $this->output->writeln("  |     '- <red>You decided to not delete the symlink automatically.</red>");
+                $this->output->writeln("  |     '- <red>You decided not to delete the symlink automatically.</red>");
 
                 return false;
             }
@@ -256,5 +281,19 @@ class UninstallCommand extends ConsoleCommand
         $this->output->writeln("  |- Checking destination...  <green>ok</green>");
 
         return true;
+    }
+
+    /**
+     * Check if package exists
+     *
+     * @param $slug
+     * @param $package
+     * @return int
+     */
+    private function packageExists($slug, $package)
+    {
+        $path = Grav::instance()['locator']->findResource($package->package_type . '://' . $slug);
+        Installer::isValidDestination($path);
+        return Installer::lastErrorCode();
     }
 }
